make Reasons page generic
